[SPARK-57335][SQL] Fix element_at internal error on Int.MinValue index

+++ sql/catalyst/src/main/java/org/apache/spark/sql/catalyst/expressions/ArrayExpressionUtils.java
@@ -39,7 +39,7 @@ public class ArrayExpressionUtils {
    * @return        the resolved 0-based position
    */
   public static int resolveArrayIndex(int length, int index, QueryContext context) {
-    if (length < Math.abs(index)) {
+    if (length < Math.abs((long) index)) {
       throw QueryExecutionErrors.invalidElementAtIndexError(index, length, context);
     }
     if (index == 0) {
